Signin Component Created - onChange and inputHandler
Following the creation of the route in App.js to /signin, the component `Signin.js` is created.
`Signin.js` is a class component that renders a form with user and password inputs and a `Sign In` Button.
Here you can clearly see how onChange works, responding to each key input it receives!
React App running without any issues!