onSubmit with axios call added to Signin.js
axios dependency added to client and required in Signin.js
if working properly, this should enable us to input a username and password, click `Sign In`, and see a status 201 returned by morgan
At this point however only status 401 (Unauthorized) is being returned, so we are debugging to find the issue
console.log(username, password) inside routes.js shows that we do have the correct username and password inside localStrategy
Turns out that everything IS working.  It is just that the user (username: frodo, password: Bodo)  was created before we wrote the code that hashed the passwords.  For some reason users created before we added the bcrypt hash are not able to be authorized.